GraphQL2/requirements.txt
aniso8601==7.0.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.3
Flask-JWT-Extended==4.5.2
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
promise==2.3
psycopg2-binary==2.9.6
PyJWT==2.0.1
python-dotenv==1.0.0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.49
Werkzeug==2.3.6
